+++ R/predict_parts.R
@@ -6,7 +6,7 @@
 #' \code{\link[iBreakDown:break_down_uncertainty]{shap}} functions from the \code{iBreakDown} package.
 #' Find information how to use this function here: \url{https://pbiecek.github.io/ema/breakDown.html}.
 #'
-#' @param explainer a model to be explained, preprocessed by the 'explain' function
+#' @param explainer a model to be explained, preprocessed by the \code{explain} function
 #' @param new_observation a new observarvation for which predictions need to be explained
 #' @param ... other parameters that will be passed to \code{iBreakDown::break_down}
 #' @param type the type of variable attributions. Either \code{shap}, \code{oscillations}, \code{break_down} or \code{break_down_interactions}.
@@ -14,36 +14,49 @@
 #' @return Depending on the \code{type} there are different classess of the resulting object.
 #' It's a data frame with calculated average response.
 #'
+#'
+#'
 #' @aliases predict_parts_break_down predict_parts predict_parts_ibreak_down predict_parts_shap
 #' @references Explanatory Model Analysis. Explore, Explain and Examine Predictive Models. \url{https://pbiecek.github.io/ema/}
-#' @examples
-#' new_dragon <- data.frame(year_of_birth = 200,
-#'      height = 80,
-#'      weight = 12.5,
-#'      scars = 0,
-#'      number_of_lost_teeth  = 5)
 #'
-#' dragon_lm_model4 <- lm(life_length ~ year_of_birth + height +
-#'                                      weight + scars + number_of_lost_teeth,
-#'                        data = dragons)
-#' dragon_lm_explainer4 <- explain(dragon_lm_model4, data = dragons, y = dragons$year_of_birth,
-#'                                 label = "model_4v")
-#' dragon_lm_predict4 <- predict_parts_break_down(dragon_lm_explainer4,
-#'                 new_observation = new_dragon)
-#' head(dragon_lm_predict4)
-#' plot(dragon_lm_predict4)
+#' @examples
+#' library(DALEX)
+#'
+#' new_dragon <- data.frame(
+#'     year_of_birth = 200,
+#'     height = 80,
+#'     weight = 12.5,
+#'     scars = 0,
+#'     number_of_lost_teeth  = 5
+#' )
+#'
+#' model_lm <- lm(life_length ~ year_of_birth + height +
+#'                weight + scars + number_of_lost_teeth,
+#'                data = dragons)
+#'
+#' explainer_lm <- explain(model_lm,
+#'                         data = dragons,
+#'                         y = dragons$year_of_birth,
+#'                         label = "model_lm")
+#'
+#' bd_lm <- predict_parts_break_down(explainer_lm, new_observation = new_dragon)
+#' head(bd_lm)
+#' plot(bd_lm)
 #'
 #' \dontrun{
 #' library("ranger")
-#' dragon_ranger_model4 <- ranger(life_length ~ year_of_birth + height +
-#'                                                weight + scars + number_of_lost_teeth,
-#'                                  data = dragons, num.trees = 50)
-#' dragon_ranger_explainer4 <- explain(dragon_ranger_model4, data = dragons, y = dragons$year_of_birth,
-#'                                 label = "model_ranger")
-#' dragon_ranger_predict4 <- predict_parts_break_down(dragon_ranger_explainer4,
-#'                                                           new_observation = new_dragon)
-#' head(dragon_ranger_predict4)
-#' plot(dragon_ranger_predict4)
+#' model_ranger <- ranger(life_length ~ year_of_birth + height +
+#'                        weight + scars + number_of_lost_teeth,
+#'                        data = dragons, num.trees = 50)
+#'
+#' explainer_ranger <- explain(model_ranger,
+#'                             data = dragons,
+#'                             y = dragons$year_of_birth,
+#'                             label = "model_ranger")
+#'
+#' bd_ranger <- predict_parts_break_down(explainer_ranger, new_observation = new_dragon)
+#' head(bd_ranger)
+#' plot(bd_ranger)
 #'}
 #'
 #' @name predict_parts
@@ -68,7 +81,10 @@ predict_parts_oscillations <- function(explainer, new_observation, ...) {
   cp <- ingredients::ceteris_paribus(explainer,
                          new_observation = new_observation,
                          ...)
-  ingredients::calculate_oscillations(cp)
+  res <- ingredients::calculate_oscillations(cp)
+
+  class(res) <- c('predict_parts', class(res))
+  res
 }
 
 #' @name predict_parts
@@ -78,9 +94,11 @@ predict_parts_break_down <- function(explainer, new_observation, ...) {
   test_explainer(explainer, has_data = TRUE, function_name = "predict_parts_break_down")
 
   # call the break_down
-  iBreakDown::break_down(explainer,
-                         new_observation = new_observation,
-                         ...)
+  res <- iBreakDown::break_down(explainer,
+                                new_observation = new_observation,
+                                ...)
+  class(res) <- c('predict_parts', class(res))
+  res
 }
 
 #' @name predict_parts
@@ -90,10 +108,12 @@ predict_parts_break_down_interactions <- function(explainer, new_observation, ..
   test_explainer(explainer, has_data = TRUE, function_name = "predict_parts_break_down_interactions")
 
   # call the break_down
-  iBreakDown::break_down(explainer,
-                         new_observation = new_observation,
-                         ...,
-                         interactions = TRUE)
+  res <- iBreakDown::break_down(explainer,
+                                new_observation = new_observation,
+                                ...,
+                                interactions = TRUE)
+  class(res) <- c('predict_parts', class(res))
+  res
 }
 
 #' @name predict_parts
@@ -103,9 +123,11 @@ predict_parts_shap <- function(explainer, new_observation, ...) {
   test_explainer(explainer, has_data = TRUE, function_name = "predict_parts_shap")
 
   # call the shap from iBreakDown
-  iBreakDown::shap(explainer,
-                         new_observation = new_observation,
-                         ...)
+  res <- iBreakDown::shap(explainer,
+                          new_observation = new_observation,
+                          ...)
+  class(res) <- c('predict_parts', class(res))
+  res
 }
 
 #' @name predict_parts